*: update proto, add indexReq, implement local region server.

+++ store/localstore/local_client.go
@@ -34,9 +34,9 @@ func (c *localClient) updateRegionInfo() {
 
 type respIterator struct {
 	client      *localClient
-	reqSent     int64
-	respGot     int64
-	concurrency int64
+	reqSent     int
+	respGot     int
+	concurrency int
 	tasks       []*task
 	responses   []*regionResponse
 	taskChan    chan *task
